requirements.txt
openai-whisper
jiwer
soundfile
tqdm
torch